refactor(aztec.js): add proofExport.js to reduce WET code
