+++ packages/aztec.js/src/proof/epoch0/BALANCED/joinSplit/verifier.js
@@ -2,7 +2,7 @@ const bn128 = require('@aztec/bn128');
 const { constants, errors } = require('@aztec/dev-utils');
 
 const Keccak = require('../../../../keccak');
-const Verifier = require('../../../verifier');
+const Verifier = require('../../../epoch0/verifier');
 
 const { ZERO_BN } = constants;
