+++ packages/aztec.js/src/proof/epoch0/MINT/mint/verifier.js
@@ -1,4 +1,4 @@
-const JoinSplitFluidVerifier = require('../../../verifier');
+const JoinSplitFluidVerifier = require('../../../epoch0/verifier');
 
 class MintVerifier66049 extends JoinSplitFluidVerifier {}
