+++ packages/aztec.js/src/proof/epoch0/UTILITY/dividend/verifier.js
@@ -4,7 +4,7 @@ const { constants, errors } = require('@aztec/dev-utils');
 const BN = require('bn.js');
 
 const Keccak = require('../../../../keccak');
-const Verifier = require('../../../verifier');
+const Verifier = require('../../verifier');
 
 const { K_MAX, ZERO_BN } = constants;
