+++ packages/aztec.js/src/proof/epoch0/joinSplitFluid/index.js
@@ -4,8 +4,8 @@ const { keccak256 } = require('web3-utils');
 
 const { inputCoder, outputCoder } = require('../../../encoder');
 const JoinSplitProof = require('../BALANCED/joinSplit');
-const { ProofType } = require('../../proof');
-const ProofUtils = require('../../utils');
+const { ProofType } = require('../proof');
+const ProofUtils = require('../utils');
 
 const { AztecError } = errors;
