+++ packages/aztec.js/src/proof/epoch0/types.js
