+++ packages/aztec.js/src/proof/epoch0/utils.js
