+++ packages/aztec.js/src/proof/proofExport.js
@@ -1,17 +1,18 @@
 /* eslint-disable func-names */
 const { errors, proofs } = require('@aztec/dev-utils');
-const BurnProof66305 = require('./epoch0/BURN/burn');
 
 const { AztecError } = errors;
 
-function ProofExport(...args) {
+const proofExport = {};
+
+proofExport.exportProof = (Proof, ...args) => {
     if (!this.epochNum) {
         this.epochNum = 1; // default epochNumber
     }
 
     switch (this.epochNum) {
         case 1:
-            return new BurnProof66305(...args);
+            return new Proof(...args);
         default:
             throw new AztecError(errors.codes.EPOCH_DOES_NOT_EXIST, {
                 message: 'The requested epoch number was greater than the latest epoch',
@@ -19,12 +20,6 @@ function ProofExport(...args) {
                 latestEpochNum: proofs.LATEST_EPOCH,
             });
     }
-}
-
-ProofExport.epoch = function(epochNum) {
-    return (...args) => {
-        return ProofExport.bind({ epochNum })(...args);
-    };
 };
 
-module.exports = ProofExport;
+module.exports = proofExport;
