+++ packages/aztec.js/test/proof/joinSplit/verifier.js
@@ -10,7 +10,7 @@ const { balancedPublicValues, mockNoteSet, randomNoteValue } = require('../../he
 const { JoinSplitProof, Proof } = require('../../../src/proof');
 const JoinSplitVerifier = require('../../../src/proof/epoch0/BALANCED/joinSplit/verifier');
 const { mockZeroJoinSplitProof } = require('../../helpers/proof');
-const ProofUtils = require('../../../src/proof/utils');
+const ProofUtils = require('../../../src/proof/epoch0/utils');
 
 describe('Join-Split Proof Verifier', () => {
     const sender = randomHex(20);
